+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 ## Change log
 ----------------------
 
+Version 5.5
+-------------
+
+ADDED:
+ 
+- new method created for create random hex string objects from a given length of characters
+
+CHANGED:
+
+- update of jobject-extensions version to 2
+- update of jcommons-lang version to 5
+- update of silly-collections version to 4.35
+- update of resourcebundle-inspector version to 3
+
 Version 5.4
 -------------